Updated ftplist.data. ftplist.py is a python script that will generate ftplist.data from mirror/Mirrors.masterlist. Feel free to rewrite the script in perl if you wish. Only the last 30 lines or so are important. The rest is 2 functions that could be used to test the URLs.
There is no sorting of the resulting list, which is the main reason I
haven't set this up to be done automatically. We can sort on languages,
but not countries, in wml (eperl). countries.wml would have to be rewritten
if we want to do this right.

CVS version numbers

english/distrib/ftplist.data: 1.6 -> 1.7 
english/distrib/ftplist.py: INITIAL -> 1.1