update translation (original revision 1.19 -> 1.20).
The new script in the original revision implements language-dependent
way of enumeration, so it is modified.  (Japanese language has different
way of listing.)  I also used a multibyte character in the script and
I am not sure whether it works well or not; I have to check this page
later.

CVS version numbers

japanese/News/weekly/index.wml: 1.9 -> 1.10