Added a long-missing web page which lists non-vulnerabilities, i.e. vulnerabilities that affect Linux or GNU/Linux and/or other vendors have reacted on them, when the dnoted Debian distribution is not vulnerable to this particular problem.
The security references (mostly CVE Ids) can be used as anchor in the
document itself to link to the particular item directly like with
<http://www.debian.org/security/nonvulns-woody#CAN-2004-0078>

This file will also contain "[reserved]" as text (or something
similar) if the Security Team discovers that this particular Debian
distribution is not vulnerable to a specific problem, but the problem
hasn't been disclosed yet.

To resolve this, it would be good if somebody would check the
generated links from time to time so the missing information could be
added.

The filename *.src was chosen so that the translation-check header
will be helpful for translators (even if it'll be ignored during page
building).

CVS version numbers

english/security/nonvulns-woody.src: INITIAL -> 1.1