Initial translation.
The file social_contract.1.1.wml is identical to social_contract.wml,
revision 1.13 in the English directory. I used the translation for that
revision, which is social_contract.wml (r1.26) in the German directory.
The only modification I did to that file was applying the fix for the
translation error stated in bug #258319.

CVS version numbers

german/social_contract.1.1.wml: INITIAL -> 1.1