Sync: Remove mentions of non-us, dead with sarge release
Part of it I didn't bother to conditionalise for sarge, so it will go live
immediatly, which should be ok, I guess. If sonmeone disagrees, feel free
to resurrect the text and wrap it in appropriate woody condfitionals.

Also, fixed up the info on mirroring using jigdo so it will point to images
that work for sarge once sarge is released (1.12)

CVS version numbers

swedish/CD/mirroring/index.wml: 1.13 -> 1.14