Patch by Jutta Wrage. Her explanations: css for the d-i css for a mail table layout (class ridgetable) as there are still layout tables and css for tables cannot be generalized. css for a div containing a quote [1] css for tag samp [2] css for normal hr (repalcing the "noshade and height 1") [3] css for a gloss dt (dt class="gloss" which renders the gloss term bold. [4]
[1] This is used forst time to replace the table at http://
www.debian.org/devel/join/newmaint but can be used with and without
image on every page.
Please note, that <blockquote><p></p></blockqote> contains the quote
and <cite></cite> the place or person, the quote is taken from. (See
HTML specification).
[2] The tags <samp> </samp> are for screen text output (defined in
HTML 4.01) this is used to replace the wrong used "blockquote" there.
Blockquotes do not get special css as they are still used for layout
purposes on many pages.
[3] In most cases the hr class=hidecss should be used to have a hr in
lynx where the graphical pages show a border. But if you want a hr
where no border is, you can use hr without class now. It will be
rendered as a 1px grey line.
[4] For glossaries there is not need to use tables! please note that
tables should be real tables (containg tabluar data how it would be
done on a sheet of paper or in a book), have a summary, table heads
(th) and a caption as title.

CVS version numbers

english/debian.css: 1.30 -> 1.31