minimal update to trademark policy to allow DFSG-free logo relicensing
As decided by the DPL, in accordance to legal advice via SPI.

This change is meant to be temporary change, to decouple the issue of
logo relicensing from the approval of a new trademark policy. Discussion
on the latter is currently ongoing on -project, see

  https://lists.debian.org/debian-project/2012/07/msg00047.html

Ideally, it will lead to a new trademark policy meant to replace
entirely the current one contained in trademark.wml

This change is not meant to affect the previous policy meaning; it just
refers more broadily to "Debian trademarks" instead of assuming there is
only one (there are more, possibly slightly different, in various
countries, etc.).

I've also removed the outer quotes, as it is no longer a true quote.

CVS version numbers

english/trademark.wml: 1.19 -> 1.20