Add recent_list_security.wml
Big refactoring of the code in recent_list.wml to
make it readable and maintaineable.
Trying to separate code that is common for all
recent list users better from the code that is
specific to the different categories of items.
This gets rid of the big unreadable if/elsif
blocks and instead replaces it with a lot of
smaller functions.

So far this only implements the security/
functionality. Other sections to follow once
it is clear that this is indeed the way
forward.

CVS version numbers

english/template/debian/recent_list_common.wml: INITIAL -> 1.1 
english/template/debian/recent_list_security.wml: INITIAL -> 1.1