Changes related to the re-ordering.
A new heading is needed for the split-out criteria.

A comment is needed to refer to the split-out criteria.

"In addition" needs to be moved to the second paragraph.

"a number of" needs to be removed to improve the grammar of the second paragraph.

CVS version numbers

english/derivatives/index.wml: 1.15 -> 1.16