Major Makefile overhauls
Add rules to generate the git revs cache in Makefile.common,
referenced as $(VCSREVCACHE)

Add an extra order-only prerequisite on $(VCSREVCACHE) for ~all rules
converting .wml to .html, so that we'll make sure we have a cache
built up-front before we start building anything.

More factorisation of various year-based Makefiles to use a common
Make.year set of rules.