Fix up logic bug in touch_translations.pl
The original logic was broken and was causing comparisons against the
wrong originals for some translated files. This looks like it was
being hidden by the lax CVS versioning code underneath, so we never
saw the bug previously.

Split the code out, correct it and document what it's actually doing
now. Fixes a lot of reported errors with git revisions not being
found...