Deal with translated pages including headers/footers/navbars etc.
If any of the included files use wml::debian::translation-check with a
version, that over-rides the version defined in the start file and
this can totally break the translation checks. The check is then made
against the start file, but using the version defined in the included
file. :-(

Fix things up to allow for this *and* check the versions of both the
start file *and* the included file:

 * Add a sym-link of translation-check.wml as
   "translation-check-inc.wml" to defeat wml's tracking of what files
   have been #use'd already. Otherwise we'll only check the included
   file and never check the start file.
 * Switch all the .inc files to use the new translation-check-inc.wml
   instead of translation-check.wml
 * Also add an extra argument "force_file" to allow the included files
   to declare their filename so *that* is checked instead of the start
   file

Quite involved, but it works...