Fix validation errors on Debian Chinese Project pages
And make some updates such as mentioning DebConf18 in Hsinchu, Taiwan,
on the News page.

Also replace HTML5 <details> tag with grey text because Debian web pages
have not migrated from HTML 4.01 to HTML5 yet, and it is not too trivial
to do so without triggering a lot of other validation errors.