Merge the translated files into the english one
Currently, output of translation jobs is not merged with the output of
the english source. The result of that is that untranslated files, such
as artwork and stylesheets, is not available for translated versions of
the site.

This is less than helpful, so add back a job which does build the
languages by using the rsync trick which we originally used for the
pages job. Give it a different name though, so that salsa doesn't try to
publish it as a pages job...